requirements.txt
flask
flask_sqlalchemy
flask_login
pandas
requests
beautifulsoup4
selenium

streamlit
plotly
scipy
scikit-optimize

gspread
google-auth
